added code and tests for REgister-PSResourceRepository
