+++ src/code/RepositorySettings.cs
@@ -10,6 +10,7 @@ using System.Management.Automation;
 using System.Xml.Linq;
 using System.Linq;
 using static System.Environment;
+using Microsoft.PowerShell.PowerShellGet.PSRepositoryItem;
 
 namespace Microsoft.PowerShell.PowerShellGet.RepositorySettings
 {
@@ -78,7 +79,7 @@ namespace Microsoft.PowerShell.PowerShellGet.RepositorySettings
         /// Returns: void
         /// </summary>
         /// <param name="sectionName"></param>
-        public void Add(string repoName, Uri repoURL, int repoPriority, bool repoTrusted)
+        public PSRespositoryItem Add(string repoName, Uri repoURL, int repoPriority, bool repoTrusted)
         {
             // Check to see if information we're trying to add to the repository is valid
             if (string.IsNullOrEmpty(repoName))
@@ -130,6 +131,11 @@ namespace Microsoft.PowerShell.PowerShellGet.RepositorySettings
 
             // Close the file
             root.Save(DefaultFullRepositoryPath);
+
+            // create PSRepositoryItem object to return
+            PSRespositoryItem repoItem = new PSRespositoryItem(repoName, repoURL, repoPriority, repoTrusted);
+
+            return repoItem;
         }
 
         /// <summary>
